Add jenkins service
For systemd services to be available, the ENTRYPOINT from the
base image (`init`) must be kept. Here, we start jenkins as a
systemd service. This, however, requires Java be installed so
the omero-install scripts have been migrated from devspace to
devslave-*. A build argument is now available to choose which
Java version should be used.